requirements.txt
google-generativeai
google-genai
pydub
SpeechRecognition
requests
beautifulsoup4
youtube-data-api
moviepy
pillow
num2words
cssutils
google-cloud-texttospeech
python-dotenv
google-api-python-client
google-auth-oauthlib 
html2image 
selenium
webdriver-manager
psutil
playwright 
playwright-stealth 